Revert "increase packet length"
This reverts commit 5ea64a63e3cdbcccd37041f88cbd93f7104423a1.

+++ SSMP/Networking/Packet/Packet.cs
@@ -83,7 +83,7 @@ internal class Packet : IPacket {
     /// </summary>
     public void WriteLength() {
         if (_buffer == null) throw new InvalidOperationException("Cannot write to Read-Only Packet");
-        var length = (int) _buffer.Count;
+        var length = (ushort) _buffer.Count;
         _buffer.Insert(0, (byte) length);
         _buffer.Insert(1, (byte) (length >> 8));
         Length = _buffer.Count;
